mt76: mt7915: fix radar detector logic
Before this patch, if AP went from ch 100 to ch 36, the radar detector
logic in the firmware was not being disabled.  This made the AP appear
to be up, but no beacons were seen on air until module reload or
reboot.

To reproduce this, I change hostapd.conf and restart hostapd.  Others
on openwrt used their UI to make changes and problem was seen, but
stil others changed channels in some other way and/or had some other
difference and could *not* reproduce it.  So, something perhaps a
bit subtle.

To fix the problem, stop depending on comparing dfs_state, store last
freq/bandwidth to detect changes in that, and streamline code that
checks to enable/disable radar detection.  And add in error checking
and dev_dbg logic so one can see what is actually happening if need
to debug this again.

NOTE:  Dumpster fire here.  Merged in some of nbd's changes so rest
of mt76 directories would compile with his dfs/cac patch.
Added some other hacks to fix up compile after earlier
merges.

Signed-off-by: Ben Greear <greearb@candelatech.com>